+++ iosHyperskillApp/iosHyperskillApp/Assets.xcassets/ProblemOfDaySolvedModal/problem-of-day-solved-modal-gems-badge.imageset/problem-of-day-solved-modal-gems-badge-light.pdf
@@ -1,9 +1,7 @@
 %PDF-1.7
 
 1 0 obj
-  << /ExtGState << /E2 << /ca 0.120000 >>
-                   /E1 << /ca 0.120000 >>
-                >> >>
+  << /ExtGState << /E1 << /ca 0.120000 >> >> >>
 endobj
 
 2 0 obj
@@ -15,120 +13,103 @@ q
 /E1 gs
 1.000000 0.000000 -0.000000 1.000000 0.000000 0.000000 cm
 0.549020 0.352941 1.000000 scn
-0.000000 24.000000 m
-0.000000 37.254833 10.745167 48.000000 24.000000 48.000000 c
-24.000000 48.000000 l
-37.254833 48.000000 48.000000 37.254833 48.000000 24.000000 c
-48.000000 24.000000 l
-48.000000 10.745167 37.254833 0.000000 24.000000 0.000000 c
-24.000000 0.000000 l
-10.745167 0.000000 0.000000 10.745167 0.000000 24.000000 c
-0.000000 24.000000 l
+0.000000 18.000000 m
+0.000000 27.941124 8.058875 36.000000 18.000000 36.000000 c
+18.000000 36.000000 l
+27.941126 36.000000 36.000000 27.941124 36.000000 18.000000 c
+36.000000 18.000000 l
+36.000000 8.058874 27.941126 0.000000 18.000000 0.000000 c
+18.000000 0.000000 l
+8.058875 0.000000 0.000000 8.058874 0.000000 18.000000 c
+0.000000 18.000000 l
 h
 f
 n
 Q
 q
-/E2 gs
-1.000000 0.000000 -0.000000 1.000000 10.000000 10.000000 cm
-0.549020 0.352941 1.000000 scn
-0.000000 14.000000 m
-0.000000 21.731987 6.268014 28.000000 14.000000 28.000000 c
-14.000000 28.000000 l
-21.731987 28.000000 28.000000 21.731987 28.000000 14.000000 c
-28.000000 14.000000 l
-28.000000 6.268013 21.731987 0.000000 14.000000 0.000000 c
-14.000000 0.000000 l
-6.268014 0.000000 0.000000 6.268013 0.000000 14.000000 c
-0.000000 14.000000 l
-h
-f
-n
-Q
-q
-1.000000 0.000000 -0.000000 1.000000 24.000008 15.327637 cm
+1.000000 0.000000 -0.000000 1.000000 18.000000 7.593201 cm
 0.854369 0.394896 0.929167 scn
-3.693969 10.647339 m
-0.000000 0.000022 l
-9.081495 10.647339 l
-3.693969 10.647339 l
+4.432763 12.776794 m
+0.000000 0.000014 l
+10.897794 12.776794 l
+4.432763 12.776794 l
 h
 f
 n
 Q
 q
-1.000000 0.000000 -0.000000 1.000000 14.918457 25.974976 cm
+1.000000 0.000000 -0.000000 1.000000 7.102173 20.369995 cm
 0.882353 0.517647 0.945098 scn
-3.757855 4.697327 m
-0.000000 -0.000004 l
-5.387563 -0.000004 l
-3.757855 4.697327 l
+4.509426 5.636780 m
+0.000000 -0.000017 l
+6.465075 -0.000017 l
+4.509426 5.636780 l
 h
 f
 n
 Q
 q
-1.000000 0.000000 -0.000000 1.000000 27.694016 25.974976 cm
+1.000000 0.000000 -0.000000 1.000000 22.432861 20.369995 cm
 0.882353 0.517647 0.945098 scn
-0.000000 -0.000004 m
-5.387522 -0.000004 l
-1.629698 4.697327 l
-0.000000 -0.000004 l
+0.000000 -0.000017 m
+6.465026 -0.000017 l
+1.955637 5.636780 l
+0.000000 -0.000017 l
 h
 f
 n
 Q
 q
-1.000000 0.000000 -0.000000 1.000000 20.306000 25.974976 cm
+1.000000 0.000000 -0.000000 1.000000 13.567261 20.369995 cm
 0.925490 0.698039 0.964706 scn
-7.387901 -0.000004 m
-3.693931 4.697327 l
-0.000000 -0.000004 l
-7.387901 -0.000004 l
+8.865481 -0.000017 m
+4.432717 5.636780 l
+0.000000 -0.000017 l
+8.865481 -0.000017 l
 h
 f
 n
 Q
 q
-1.000000 0.000000 -0.000000 1.000000 18.676476 25.974976 cm
+1.000000 0.000000 -0.000000 1.000000 11.611816 20.369995 cm
 0.854369 0.394896 0.929167 scn
-5.323638 4.697327 m
-0.000000 4.697327 l
-1.629707 -0.000004 l
-5.323638 4.697327 l
+6.388366 5.636780 m
+0.000000 5.636780 l
+1.955649 -0.000017 l
+6.388366 5.636780 l
 h
 f
 n
 Q
 q
-1.000000 0.000000 -0.000000 1.000000 24.000008 25.974976 cm
+1.000000 0.000000 -0.000000 1.000000 18.000000 20.369995 cm
 0.854369 0.394896 0.929167 scn
-5.323668 4.697327 m
-0.000000 4.697327 l
-3.693969 -0.000004 l
-5.323668 4.697327 l
+6.388401 5.636780 m
+0.000000 5.636780 l
+4.432763 -0.000017 l
+6.388401 5.636780 l
 h
 f
 n
 Q
 q
-1.000000 0.000000 -0.000000 1.000000 20.306000 15.327637 cm
+1.000000 0.000000 -0.000000 1.000000 13.567261 7.593201 cm
 0.882353 0.517647 0.945098 scn
-0.000000 10.647339 m
-3.693931 0.000022 l
-7.387901 10.647339 l
-0.000000 10.647339 l
+0.000000 12.776794 m
+4.432718 0.000014 l
+8.865481 12.776794 l
+0.000000 12.776794 l
 h
 f
 n
 Q
 q
-1.000000 0.000000 -0.000000 1.000000 14.918457 15.327637 cm
+1.000000 0.000000 -0.000000 1.000000 7.102173 7.593201 cm
 0.854369 0.394896 0.929167 scn
-0.000000 10.647339 m
-9.081494 0.000022 l
-5.387563 10.647339 l
-0.000000 10.647339 l
+0.000000 12.776794 m
+10.897794 0.000014 l
+6.465076 12.776794 l
+0.000000 12.776794 l
 h
 f
 n
@@ -138,13 +119,13 @@ endstream
 endobj
 
 3 0 obj
-  2403
+  1943
 endobj
 
 4 0 obj
   << /Annots []
      /Type /Page
-     /MediaBox [ 0.000000 0.000000 48.000000 48.000000 ]
+     /MediaBox [ 0.000000 0.000000 36.000000 36.000000 ]
      /Resources 1 0 R
      /Contents 2 0 R
      /Parent 5 0 R
@@ -168,16 +149,16 @@ xref
 0 7
 0000000000 65535 f
 0000000010 00000 n
-0000000132 00000 n
-0000002591 00000 n
-0000002614 00000 n
-0000002787 00000 n
-0000002861 00000 n
+0000000074 00000 n
+0000002073 00000 n
+0000002096 00000 n
+0000002269 00000 n
+0000002343 00000 n
 trailer
 << /ID [ (some) (id) ]
    /Root 6 0 R
    /Size 7
 >>
 startxref
-2920
+2402
 %%EOF
