+++ iosHyperskillApp/iosHyperskillApp/Sources/Models/Constants/Images/Images.swift
@@ -55,13 +55,11 @@ enum Images {
             static let helpfulReaction = "step_quiz_hints_helpful_reaction"
             static let unhelpfulReaction = "step_quiz_hints_unhelpful_reaction"
         }
-    }
 
-    // MARK: - ProblemOfDaySolvedModal -
-
-    enum ProblemOfDaySolvedModal {
-        static let book = "problem-of-day-solved-modal-book"
-        static let gemsBadge = "problem-of-day-solved-modal-gems-badge"
+        enum ProblemOfDaySolvedModal {
+            static let book = "problem-of-day-solved-modal-book"
+            static let gemsBadge = "problem-of-day-solved-modal-gems-badge"
+        }
     }
 
     // MARK: - Track -
