+++ shared/src/commonMain/resources/MR/base/strings.xml
@@ -63,12 +63,10 @@
     <string name="step_quiz_checking_button_text">Checking</string>
     <string name="step_quiz_discussions_button_text">Show discussions</string>
     <string name="step_quiz_unsupported_quiz_text">Current quiz type is not supported in mobile app yet</string>
-
-    <!-- Problem of day solved modal -->
-    <string name="problem_of_day_solved_modal_title">Problem of the day\nsolved!</string>
-    <string name="problem_of_day_solved_modal_text">Solve another one tomorrow to get more</string>
-    <string name="problem_of_day_solved_modal_total_gems">Total gems</string>
-    <string name="problem_of_day_solved_modal_go_back_button_text">Go back</string>
+    <string name="step_quiz_problem_of_day_solved_modal_title">Problem of the day\nsolved!</string>
+    <string name="step_quiz_problem_of_day_solved_modal_text">Solve another one tomorrow to get more</string>
+    <string name="step_quiz_problem_of_day_solved_modal_total_gems">Total gems</string>
+    <string name="step_quiz_problem_of_day_solved_modal_go_back_button_text">Go back</string>
 
     <!-- Step quiz hint -->
     <string name="step_quiz_hints_show_button_text">See hint</string>
